+++ src/main/java/ru/mels/tasktracking/dto/TaskBoardRequestDto.java
@@ -1,11 +1,7 @@
 package ru.mels.tasktracking.dto;
 
-/**
- * @author Meleshkin Alexandr
- * @since 07.03.2022
- */
 public class TaskBoardRequestDto {
-    Integer id;
+    Long id;
     String title;
 
     public String getTitle() {
