+++ src/main/java/ru/mels/tasktracking/dto/TaskBoardResponseDto.java
@@ -1,19 +1,15 @@
 package ru.mels.tasktracking.dto;
 
-/**
- * @author Meleshkin Alexandr
- * @since 07.03.2022
- */
 public class TaskBoardResponseDto {
-    Integer id;
+    Long id;
     String title;
-    Integer projectId;
+    Long projectId;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -25,11 +21,11 @@ public class TaskBoardResponseDto {
         this.title = title;
     }
 
-    public Integer getProjectId() {
+    public Long getProjectId() {
         return projectId;
     }
 
-    public void setProjectId(Integer projectId) {
+    public void setProjectId(Long projectId) {
         this.projectId = projectId;
     }
 }
