+++ src/main/java/ru/mels/tasktracking/exception/ReleaseNotFoundException.java
@@ -1,11 +1,7 @@
 package ru.mels.tasktracking.exception;
 
-/**
- * @author Meleshkin Alexandr
- * @since 07.03.2022
- */
 public class ReleaseNotFoundException extends NotFoundElementException{
-    public ReleaseNotFoundException(Integer id) {
+    public ReleaseNotFoundException(Long id) {
         super("Release with id - " + id + " is not found");
     }
 }
