+++ src/main/java/ru/mels/tasktracking/repository/TaskRepository.java
@@ -1,11 +1,12 @@
 package ru.mels.tasktracking.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 import ru.mels.tasktracking.entity.Task;
+import ru.mels.tasktracking.enums.TaskStatus;
 
-/**
- * @author Meleshkin Alexandr
- * @since 07.03.2022
- */
-public interface TaskRepository extends JpaRepository<Task, Integer> {
+import java.util.List;
+
+public interface TaskRepository extends JpaRepository<Task, Long> {
+
 }
